[Rsbuild] Serve Rsbuild through unplugin's createRsbuildPlugin
unplugin 3 adds Rsbuild support, so the hand-written native RsbuildPlugin — a
workaround for unplugin 2.x having none — is replaced by the shared
unpluginFactory. The Stimulus virtual module moves to universal resolveId/load
hooks for both bundlers, dropping the VirtualModulesPlugin +
NormalModuleReplacementPlugin Rspack hack. @rsbuild/core is imported lazily
inside the rsbuild setup hook so it stays out of the Vite bundle.

Also fix the dev-server host: HMR, lazy compilation, and async chunks now
target the host the dev server binds to (default localhost) instead of a
hardcoded 127.0.0.1, which is not listening when Rsbuild binds to ::1.
