+++ AGENTS.md
@@ -39,12 +39,12 @@ Bundler-agnostic core + per-bundler adapters, all under `assets/src/`:
 
 - `assets/src/core/` — pure, no bundler imports: `options.ts` (`normalizeOptions` + CDN guard + `resolvePublicPath`), `dev-server.ts` (`resolveDevOrigin`), `format.ts` (`buildEntrypoints`/`buildManifest` in the frozen v1 format), `emit.ts` (`writeSymfonyFiles`).
 - `assets/src/collectors/` — turn a bundler's output into the shared `NormalizedGraph`: `vite.ts` (`bundleToGraph` from the Rollup bundle in build, `configToDevGraph` from the resolved config in serve) and `rspack.ts` (`statsToGraph` from the Rspack stats JSON).
-- `assets/src/index.ts` — the `unpluginFactory` (Vite only now) + `createUnplugin` default export. Its `vite` hooks call the collectors + core: `config()` sets `base`/`outDir` and disables Vite's own manifest/publicDir copy; `generateBundle` emits the two files on build; `configureServer` writes the dev-flavoured files pointing at the dev-server origin.
+- `assets/src/index.ts` — the `unpluginFactory` (Vite + Rsbuild) + `createUnplugin` default export. Universal `resolveId`/`load` serve the Stimulus virtual module for both bundlers. Its `vite` hooks call the collectors + core: `config()` sets `base`/`outDir` and disables Vite's own manifest/publicDir copy; `generateBundle` emits the two files on build; `configureServer` writes the dev-flavoured files pointing at the dev-server origin. Its `rsbuild` hook's `setup(api)` does the same job for Rsbuild — `api.modifyRsbuildConfig` forces `tools.htmlPlugin = false` (no per-entry HTML), disables the public-dir copy (output lives under `public/build`), and sets the output paths + dev origin; `api.onAfterCreateCompiler` taps `compiler.hooks.done` to run `statsToGraph` + core. `@rsbuild/core` is an optional peer, so it is `await import()`ed lazily inside `setup` (never a static import) to keep it out of the Vite bundle.
 - `assets/src/vite.ts` — one-line unplugin adapter `createVitePlugin(unpluginFactory)` (`@symfony/reprise/vite`).
-- `assets/src/rsbuild.ts` — a **hand-written native `RsbuildPlugin`** (default export, `@symfony/reprise/rsbuild`), **not** an unplugin adapter. unplugin has no `createRsbuildPlugin`, and the Symfony integration needs Rsbuild-config-level control a raw Rspack plugin can't reach: `api.modifyRsbuildConfig` forces `tools.htmlPlugin = false` (no per-entry HTML) and disables the public-dir copy (output lives under `public/build`), and sets the output paths + dev origin; `api.onAfterCreateCompiler` taps `compiler.hooks.done` to run `statsToGraph` + core. It reuses the same core as the Vite path.
+- `assets/src/rsbuild.ts` — one-line unplugin adapter `createRsbuildPlugin(unpluginFactory)` (`@symfony/reprise/rsbuild`). unplugin's Rsbuild adapter forwards the real Rsbuild plugin `api` to the factory's `rsbuild.setup` hook (so full config-level control is preserved) and injects the raw plugin into the Rspack config via `api.modifyRspackConfig`, which is how the universal `resolveId`/`load` reach Rspack.
 - `assets/src/types.ts` — public `Options` (`outputPath`, `publicPath`, `manifestKeyPrefix`, `devServerOrigin`) + the frozen `EntrypointsJson`/`ManifestJson`/`EntryFiles` shapes (`js`/`css`/`preload`/`dynamic`).
 
-unplugin still earns its place for Vite and (upcoming) the Stimulus virtual module (universal `resolveId`/`load`, cross-bundler). Rspack is served exclusively through the native Rsbuild adapter — the raw Rspack unplugin adapter was dropped (Rsbuild is the supported Rspack layer).
+Both bundlers go through the one `unpluginFactory`: Vite via `createVitePlugin`, Rsbuild via `createRsbuildPlugin` (unplugin ≥3 added Rsbuild support — the earlier hand-written native `RsbuildPlugin` existed only because unplugin 2.x had none). The shared factory is what lets the Stimulus virtual module be a single universal `resolveId`/`load` implementation instead of per-bundler plumbing. The raw Rspack unplugin adapter is not used — Rsbuild is the supported Rspack layer.
 
 ## The Symfony integration contract (the core of this project)
