+++ CHANGELOG.md
@@ -1,0 +1,11 @@
+# CHANGELOG
+
+## 0.2.0
+
+- Update Unplugin from ^2.3.4 to ^3.3.0
+- Rsbuild support now runs on unplugin's `createRsbuildPlugin`
+- Fix the Rsbuild dev server to target `localhost` instead of a hardcoded `127.0.0.1`
+
+## 0.1.0
+
+- Initial release
