+++ assets/package.json
@@ -73,7 +73,7 @@
     }
   },
   "dependencies": {
-    "unplugin": "^2.3.4"
+    "unplugin": "^3.3.0"
   },
   "devDependencies": {
     "@hotwired/stimulus": "^3.2.2",
